# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hotel Booking API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Users
...
  /bookings:
    get:
      operationId: listAllBookings
      tags: [bookings]
      summary: List all bookings (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookingWithDetails"

    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a new booking
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Room not available or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get a booking by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingWithDetails"
        "404":
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bookings/{id}/cancel:
    post:
      operationId: cancelBooking
      tags: [bookings]
      summary: Cancel a booking
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Cannot cancel this booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bookings/{id}/pay:
    post:
      operationId: payBooking
      tags: [bookings]
      summary: Process payment for a booking (fake)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayBookingBody"
      responses:
        "200":
          description: Payment processed, booking confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Payment failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Reviews
...
  /favorites:
    get:
      operationId: getFavorites
      tags: [favorites]
      summary: Get user favorites
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of favorite hotels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Hotel"
  /favorites/{hotelId}:
    post:
      operationId: addFavorite
      tags: [favorites]
      summary: Add hotel to favorites
      security:
        - bearerAuth: []
      parameters:
        - name: hotelId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Added to favorites
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Favorite"

    delete:
      operationId: removeFavorite
      tags: [favorites]
      summary: Remove hotel from favorites
      security:
        - bearerAuth: []
      parameters:
        - name: hotelId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed from favorites
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    RegisterBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        name:
          type: string
      required:
        - email
        - password
        - name
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [USER, ADMIN]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - role
        - createdAt
...
    Hotel:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        city:
          type: string
        address:
          type: string
        rating:
          type: number
        stars:
          type: integer
        amenities:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
        reviewCount:
          type: integer
        minPrice:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - city
        - address
        - rating
        - stars
        - amenities
        - images
        - reviewCount
        - createdAt
...
    Room:
      type: object
      properties:
        id:
          type: integer
        hotelId:
          type: integer
        type:
          type: string
          enum: [single, double, deluxe, suite]
        price:
          type: number
        guests:
          type: integer
        description:
          type: string
        images:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - hotelId
        - type
        - price
        - guests
        - description
        - images
        - createdAt
...
    Booking:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        roomId:
          type: integer
        checkIn:
          type: string
          format: date-time
        checkOut:
          type: string
          format: date-time
        totalPrice:
          type: number
        status:
          type: string
          enum: [pending, confirmed, cancelled]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - roomId
        - checkIn
        - checkOut
        - totalPrice
        - status
        - createdAt
    BookingWithDetails:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        roomId:
          type: integer
        checkIn:
          type: string
          format: date-time
        checkOut:
          type: string
          format: date-time
        totalPrice:
          type: number
        status:
          type: string
          enum: [pending, confirmed, cancelled]
        createdAt:
          type: string
          format: date-time
        room:
          $ref: "#/components/schemas/Room"
        hotel:
          $ref: "#/components/schemas/Hotel"
      required:
        - id
        - userId
        - roomId
        - checkIn
        - checkOut
        - totalPrice
        - status
        - createdAt
        - room
        - hotel
    CreateBookingBody:
      type: object
      properties:
        roomId:
          type: integer
        checkIn:
          type: string
          format: date-time
        checkOut:
          type: string
          format: date-time
      required:
        - roomId
        - checkIn
        - checkOut
    PayBookingBody:
      type: object
      properties:
        cardNumber:
          type: string
        expiryDate:
          type: string
        cvv:
          type: string
      required:
        - cardNumber
        - expiryDate
        - cvv
...
    Favorite:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        hotelId:
          type: integer
      required:
        - id
        - userId
        - hotelId